backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
sqlalchemy
python-jose[cryptography]
passlib[bcrypt]
python-dotenv
pydantic-settings
alembic
psycopg2-binary
bcrypt==4.0.1